pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "hapm"
authors = [
    {name = "Mikhael Khrustik", email = "misha@myrt.co"},
]
description = "Home Assistant package management applications"
readme = "README.md"
requires-python = ">=3.12"
keywords = ["cli", "home-assistant", "package-manager"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "ruamel.yaml>=0.18.1",
    "arrrgs>=3.0.0",
    "PyGithub>=2.5.0",
    "gitpython>=3.1.44",
    "requests>=2.32.3",
    "urllib3>=2.3.0"
]
dynamic = ["version"]

[project.scripts]
hapm = "hapm.cli:start"

[tool.setuptools.dynamic]
version = {file = [".version"]}

[tool.ruff]
lint.ignore = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
lint.select = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
